pyproject.toml
[project]
name = "arduinoapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "pyserial>=3.5",
    "uvicorn>=0.41.0",
]

[project.scripts]
arduino-cli = "app.cli:main"
